Layout Position and Window functionality
From class on Thursday and today, I've been able to have a static header and .side-nav, which also contains the footer. The class .content-block serves as a dynamic window to view the content within .inner-content-block. I plan on implementing .game as one image with a brief description of each game with a link to each respective website.